python/day-4-capstone/loanhub/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pytest==8.3.4
httpx==0.28.1
anyio==4.7.0
